관리자 1대1 문의 댓글 & 회원 목록 조회 API 구현

+++ src/main/java/com/project/Real_Moment/domain/repository/GradeRepository.java
@@ -3,5 +3,5 @@ package com.project.Real_Moment.domain.repository;
 import com.project.Real_Moment.domain.entity.Grade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeRepository extends JpaRepository<Grade, Integer> {
+public interface GradeRepository extends JpaRepository<Grade, Long> {
 }
